requirements.txt
Pillow>=8.0.1
PyYAML>=5.1
albumentations>=0.5.2
antlr4-python3-runtime>=4.8
comet_ml>=3.1.6
hydra-core==1.0.4
imageio>=2.9.0
imgaug>=0.4.0
importlib-resources>=3.3.0
matplotlib>=3.3.3
nni>=1.9
numpy>=1.19.4
omegaconf>=2.0.5
onnx>=1.8.0
opencv_python_headless>=4.2.0.34
pandas>=1.1.4
pycocotools>=2.0.2
pytorch-lightning>=1.0.8
rich>=9.3.0
scikit-image>= 0.17.2
scikit-learn>=0.23.2
tensorboard>= 2.3.0
torch>=1.7.0
torchvision==0.8.1
tqdm>=4.54.0